+++ build-ant.ps1
@@ -1,2 +1,0 @@
-Set-Location "C:\NicoCache_nl"
-ant version extract jar